SAA-1256: Remove create appointment set from appointment series service
